requirements.txt
torch >= 1.4.0
torchvision
pillow
numpy
tqdm
timm
easydict
pyyaml == 5.4.1
isort
black
pre-commit
